requirements.txt
pandas>=2.0
psycopg2-binary>=2.9
pyyaml>=6.0